Simulator ve controller'lar beraber geliştirildi ve yetkilendirme yapıldı.

+++ Controllers/accountController.py
@@ -1,5 +1,6 @@
 from Network.bottle import *
 from Business.AccountBusiness import LoginControl
+from Network.Security import *
 
 account = 0
 print("accountController")
@@ -7,28 +8,25 @@ print("accountController")
 
 @post('/login')
 def login():
-    username = request.json.get('username')
-    password = request.json.get('password')
-    print("Kullanıcı adı şifre:", username, password)
-    if LoginControl(username, password):
-        response.set_cookie("account", username, secret='some-secret-key')
-        print("Başarılı giriş!")
-        return ""
+    if IsAllow(request, Roller.Misafir):
+        username = request.json.get('username')
+        password = request.json.get('password')
+        print("Kullanıcı adı şifre:", username, password)
+        if LoginControl(username, password):
+            response.set_cookie("account", username, secret='some-secret-key')
+            print("Başarılı giriş!")
+            return ""
+        else:
+            print("Hatalı giriş!")
+            return "Kullanıcı adı veya parola yanlış!"
     else:
-        print("Hatalı giriş!")
-        return "Kullanıcı adı veya parola yanlış!"
+        abort(code=500, text=ErrorText.get('500'))
 
 
 @route('/logout')
 def log_out():
-    print("Çıkış yapıldı!")
-    response.delete_cookie("account")
-
-
-@route('/restricted')
-def restricted_area():
-    username = request.get_cookie("account", secret='some-secret-key')
-    if username:
-        return ""
+    if IsAllow(request, Roller.TumHesaplar):
+        print("Çıkış yapıldı!")
+        response.delete_cookie("account")
     else:
-        return "Yetkisiz erişim!"
+        abort(code=500, text=ErrorText.get('500'))
